requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.7.0
azure-functions-durable==1.0.0
